Delegate input to #[proc_macro] at Rust 1.45+
